refactor iOS part
