+++ iosHyperskillApp/iosHyperskillApp/Sources/Models/Constants/Strings.swift
@@ -74,6 +74,11 @@ enum Strings {
     enum Step {
         static let startPracticing = sharedStrings.step_start_practicing_text.localized()
         static let theory = sharedStrings.step_theory_text.localized()
+
+        enum RequestDailyNotificationsAlert {
+            static let title = sharedStrings.after_daily_step_completed_dialog_title.localized()
+            static let text = sharedStrings.after_daily_step_completed_dialog_text.localized()
+        }
     }
 
     // MARK: - StepQuiz -
@@ -90,6 +95,11 @@ enum Strings {
         static let checkingButton = sharedStrings.step_quiz_checking_button_text.localized()
         static let discussionsButton = sharedStrings.step_quiz_discussions_button_text.localized()
         static let unsupportedText = sharedStrings.step_quiz_unsupported_quiz_text.localized()
+
+        enum ResetCodeAlert {
+            static let title = sharedStrings.reset_code_dialog_title.localized()
+            static let text = sharedStrings.reset_code_dialog_explanation.localized()
+        }
 
         enum Hints {
             static let showButton = sharedStrings.step_quiz_hints_show_button_text.localized()
