+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/StepQuiz/StepQuizAssembly.swift
@@ -24,8 +24,6 @@ final class StepQuizAssembly: Assembly {
         let commonComponent = AppGraphBridge.sharedAppGraph.commonComponent
         let stepQuizComponent = AppGraphBridge.sharedAppGraph.buildStepQuizComponent(stepRoute: self.stepRoute)
 
-        let notificationComponent = AppGraphBridge.sharedAppGraph.buildNotificationComponent()
-
         let viewDataMapper = StepQuizViewDataMapper(
             stepQuizStatsTextMapper: stepQuizComponent.stepQuizStatsTextMapper,
             stepQuizTitleMapper: stepQuizComponent.stepQuizTitleMapper
@@ -41,14 +39,7 @@ final class StepQuizAssembly: Assembly {
             moduleOutput: moduleOutput,
             provideModuleInputCallback: provideModuleInputCallback,
             viewDataMapper: viewDataMapper,
-            userPermissionRequestTextMapper: StepQuizUserPermissionRequestTextMapper(
-                resourceProvider: commonComponent.resourceProvider
-            ),
             problemsLimitViewStateMapper: problemsLimitComponent.problemsLimitViewStateMapper,
-            notificationService: NotificationsService(
-                notificationInteractor: notificationComponent.notificationInteractor
-            ),
-            notificationsRegistrationService: .shared,
             feature: stepQuizComponent.stepQuizFeature
         )
