+++ shared/src/commonMain/kotlin/org/hyperskill/app/step_quiz/injection/StepQuizComponentImpl.kt
@@ -13,7 +13,6 @@ import org.hyperskill.app.step_quiz.presentation.StepQuizFeature
 import org.hyperskill.app.step_quiz.remote.AttemptRemoteDataSourceImpl
 import org.hyperskill.app.step_quiz.view.mapper.StepQuizStatsTextMapper
 import org.hyperskill.app.step_quiz.view.mapper.StepQuizTitleMapper
-import org.hyperskill.app.step_quiz.view.mapper.StepQuizUserPermissionRequestTextMapper
 import ru.nobird.app.presentation.redux.feature.Feature
 
 class StepQuizComponentImpl(
@@ -31,9 +30,6 @@ class StepQuizComponentImpl(
 
     override val stepQuizTitleMapper: StepQuizTitleMapper
         get() = StepQuizTitleMapper(appGraph.commonComponent.resourceProvider)
-
-    override val stepQuizUserPermissionRequestTextMapper: StepQuizUserPermissionRequestTextMapper
-        get() = StepQuizUserPermissionRequestTextMapper(appGraph.commonComponent.resourceProvider)
 
     private val attemptRemoteDataSource: AttemptRemoteDataSource = AttemptRemoteDataSourceImpl(
         appGraph.networkComponent.authorizedHttpClient
